More AI QOL Stuff (#1761)
Needed for antagonistic AI gameplay. Currently any variant of
"Antagonistic AI" is extremely limited in their options for harming
crew. Traditionally SS13 AIs would have had a larger variety of options
for Harming Crew, the most iconic of which was by Plasmaflooding. With
this change, AI can "Assist" in setting up Atmos by turning on pumps and
configuring gas mixers to provide the station with breathable air. Or a
more murderous AI can "Assist" by configuring the air supply to include
plasma.

Also whitelists the Emitters and Singularity Controller for the same
purpose.

The Research Director now starts with a spare AI Upload circuitboard, as
well as 4 basic AI law boards in his locker.

I suggest writing it in your server rules whether or not plasmaflooding
is acceptable for MALF AI.

:cl:
- add: Station AI can now helpfully assist by configuring the station's
Atmos system, in order to provide crew with breathable oxygen, and
definitely not anything malicious. You can also now ask the AI to assist
with the station's power by having it turn on or off the supermatter for
safe operation.
- add: The Mystagogue's locker now starts with an AI Upload
circuitboard, as well as 4 basic AI lawset chips in his locker.